+++ packages/asc-ui/src/components/Footer/FooterSection.tsx
@@ -1,6 +1,6 @@
 import { ChevronDown } from '@amsterdam/asc-assets'
 import { FunctionComponent } from 'react'
-import { ShowHideTypes } from '../../utils'
+import { Theme } from '../../types'
 import Hidden from '../Hidden'
 import Icon from '../Icon/Icon'
 import { ToggleHandlerProps } from '../Toggle'
@@ -14,7 +14,23 @@ import {
 
 type FooterContentProps = {
   title?: string
-} & ShowHideTypes
+  ssr?: boolean
+  toggleAt?: keyof Theme.BreakpointsInterface
+  /**
+   * @deprecated Use toggleAt instead.
+   */
+  hideAt?: keyof Theme.BreakpointsInterface
+  /**
+   * @deprecated Use toggleAt instead.
+   */
+  showAt?: keyof Theme.BreakpointsInterface
+}
+
+type ConditionalWrapperProps = {
+  condition: boolean
+  wrapper: (wrapperChildren: JSX.Element) => JSX.Element
+  children: JSX.Element
+}
 
 const ToggleFooterHeader: FunctionComponent<ToggleHandlerProps> = ({
   open,
@@ -31,29 +47,53 @@ const ToggleFooterHeader: FunctionComponent<ToggleHandlerProps> = ({
   </StyledButton>
 )
 
+const ConditionalWrapper: FunctionComponent<ConditionalWrapperProps> = ({
+  condition,
+  wrapper,
+  children,
+}) => (condition ? wrapper(children) : children)
+
 const FooterSection: FunctionComponent<FooterContentProps> = ({
   title,
+  ssr,
+  toggleAt = 'tabletM',
+  showAt,
+  hideAt,
   children,
-  ...otherProps
-}) => (
-  <>
-    <Hidden minBreakpoint="tabletM">
-      <StyledFooterToggle
-        ToggleHandler={ToggleFooterHeader}
-        title={title}
-        align="left"
-        {...otherProps}
+}) => {
+  const breakpoint = showAt || hideAt || toggleAt
+
+  return (
+    <>
+      <ConditionalWrapper
+        condition={!ssr}
+        wrapper={(wrapperChildren) => (
+          <Hidden minBreakpoint={breakpoint}>{wrapperChildren}</Hidden>
+        )}
       >
-        <FooterContentWrapper>{children}</FooterContentWrapper>
-      </StyledFooterToggle>
-    </Hidden>
-    <Hidden maxBreakpoint="tabletM">
-      <FooterContentWrapper>
-        <FooterHeading>{title}</FooterHeading>
-        {children}
-      </FooterContentWrapper>
-    </Hidden>
-  </>
-)
+        <StyledFooterToggle
+          ToggleHandler={ToggleFooterHeader}
+          title={title}
+          align="left"
+          ssr={ssr}
+          breakpoint={breakpoint}
+        >
+          <FooterContentWrapper>{children}</FooterContentWrapper>
+        </StyledFooterToggle>
+      </ConditionalWrapper>
+      <ConditionalWrapper
+        condition={!ssr}
+        wrapper={(wrapperChildren) => (
+          <Hidden maxBreakpoint={breakpoint}>{wrapperChildren}</Hidden>
+        )}
+      >
+        <FooterContentWrapper ssr={ssr} breakpoint={breakpoint}>
+          <FooterHeading>{title}</FooterHeading>
+          {children}
+        </FooterContentWrapper>
+      </ConditionalWrapper>
+    </>
+  )
+}
 
 export default FooterSection